backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
google-generativeai
groq
python-multipart
